heartattack/requirements.txt
pandas
sklearn
pyyaml
numpy
joblib
